Add an abstract base class for long-running, single-instance tasks.
See single_instance_task.py.  This takes care of checking the lock/
setting the cache key when the task is called, removing the cache key
on error or task completion, and finding any running instance of the
task.  Now we don't need to muck around with the cache directly in
views.py.

Also fixed a bug with views.py:task_info, when the completed task
returns any non-serializable objects.