Rewrite publish.py, again.
Publish only one RSS feed containing all exports as well as keywords.

Preserve sanity by making export types into classes. TODO: should the
export methods be moved in there too? Should it be a model with
foreign key referencing Track?

Publish two files for each export: the content, with a timestamp in
the filename/URL; and a symlink to the real file, whose name does not
change.  This makes it easier to import the updated versions into
Drupal automatically.